New system: Resources.
Will contains all resources. Like brush alphas, textures and something else.
I added S to all class names because there is already Resources namespace in Unity. S is the first letter in the name of the program.
SResource is a base class for all resources. Contains name and preview texture property. Children of it will override preview texture.
STexture2DResource. The name spoke for itself. It is a resource, that contains a texture2D.
SResources. Singleton. Handles loading of resources from any given path. Automatically determinates what type of resource does a file contains and loads its propertly. After loading adds it to the Resources (this is property) array.